Backend-2/backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-colorfield==0.9.0
django-extra-fields==3.0.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-extra-fields==3.5.0
filetype==1.2.0
flake8==5.0.4
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
pip==23.1.2
pycparser==2.21
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2023.3
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
setuptools==49.2.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
gunicorn==20.0.4
psycopg2-binary==2.8.6
atomicwrites==1.4.1
zipp==3.15.0
django-cors-headers==3.13.0
